Use explicit rather than implicit conversion on those lines generating warnings for such.

+++ src/Avalanche/Credential.cpp
@@ -13,7 +13,7 @@ using namespace TW::Avalanche;
 void Credential::encode(Data& data) const {
     encode32BE(TypeID, data);
     // do not sort signatures, they must match input signature indices
-    encode32BE(Signatures.size(), data);
+    encode32BE(static_cast<uint32_t>(Signatures.size()), data);
     for (auto sig : Signatures) {
         for (auto byte : sig) {
             data.push_back(byte);
